[Fix]Splitted request.go
