+++ command/auth.go
@@ -1,0 +1,15 @@
+package command
+
+import "isso0424/spotify_CLI/auth"
+
+// Execute is excution command function.
+func (cmd refresh) Execute(token *string) error {
+	tokenPtr, err := auth.GetToken()
+	if err != nil {
+		return err
+	}
+
+	*token = *tokenPtr
+
+	return nil
+}
